requirements.txt
anyio==4.3.0
APScheduler==3.6.3
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
prettytable==3.10.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==13.13
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
telegram==0.0.1
tornado==6.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
wikipedia==1.4.0
